feat: switch to using mixins
